git-svn-id: https://oceanai.mit.edu/svn/moos-ivp-aro/trunk@7474 7b6de600-d9ab-45c9-a1ca-e8145d3c6daf

+++ ivp/src/app_alogview/EncounterViewer.h
@@ -76,7 +76,7 @@ class EncounterViewer : public Fl_Gl_Window
   double getCurrIndexTime() const;
   double getCurrIndexCPA() const;
   double getCurrIndexEFF() const;
-  double getCurrIndexID() const;
+  int    getCurrIndexID() const;
   std::string getCurrIndexContact() const;
   
   std::string getTotalEncounters() const;
